+++ extensions/broker/src/main/java/de/sovity/extension/broker/sender/RegisterResourceRequestSender.java
@@ -14,10 +14,14 @@
 package de.sovity.extension.broker.sender;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.node.ObjectNode;
+import de.fraunhofer.iais.eis.ConnectorEndpointBuilder;
 import de.fraunhofer.iais.eis.DynamicAttributeToken;
+import de.fraunhofer.iais.eis.IANAMediaTypeBuilder;
 import de.fraunhofer.iais.eis.Language;
 import de.fraunhofer.iais.eis.Message;
 import de.fraunhofer.iais.eis.MessageProcessedNotificationMessageImpl;
+import de.fraunhofer.iais.eis.RepresentationBuilder;
 import de.fraunhofer.iais.eis.ResourceBuilder;
 import de.fraunhofer.iais.eis.ResourceUpdateMessageBuilder;
 import de.fraunhofer.iais.eis.util.TypedLiteral;
@@ -28,6 +32,9 @@ import org.eclipse.edc.protocol.ids.api.multipart.dispatcher.sender.response.Mul
 import org.eclipse.edc.protocol.ids.spi.domain.IdsConstants;
 import org.eclipse.edc.protocol.ids.util.CalendarUtil;
 
+import java.net.URI;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 import static org.eclipse.edc.protocol.ids.api.multipart.dispatcher.sender.util.ResponseUtil.parseMultipartStringResponse;
@@ -46,42 +53,159 @@ public class RegisterResourceRequestSender implements MultipartSenderDelegate<Re
     public Message buildMessageHeader(RegisterResourceMessage registerResourceMessage,
                                       DynamicAttributeToken token) throws Exception {
         return new ResourceUpdateMessageBuilder(registerResourceMessage.affectedResourceUri())
-                        ._modelVersion_(IdsConstants.INFORMATION_MODEL_VERSION)
-                        ._issued_(CalendarUtil.gregorianNow())
-                        ._securityToken_(token)
-                        ._issuerConnector_(registerResourceMessage.connectorBaseUrl())
-                        ._senderAgent_(registerResourceMessage.connectorBaseUrl())
-                        ._affectedResource_(registerResourceMessage.affectedResourceUri())
-                        .build();
+                ._modelVersion_(IdsConstants.INFORMATION_MODEL_VERSION)
+                ._issued_(CalendarUtil.gregorianNow())
+                ._securityToken_(token)
+                ._issuerConnector_(registerResourceMessage.connectorBaseUrl())
+                ._senderAgent_(registerResourceMessage.connectorBaseUrl())
+                ._affectedResource_(registerResourceMessage.affectedResourceUri())
+                .build();
     }
 
     @Override
     public String buildMessagePayload(RegisterResourceMessage registerResourceMessage) throws Exception {
         var assetTitle = getAssetTitle(registerResourceMessage);
         var assetDescription = getAssetDescription(registerResourceMessage);
+        var language = Arrays.stream(Language.values())
+                .filter(l -> l.getId().toString().equals(getAssetLanguage(registerResourceMessage)))
+                .findFirst()
+                .orElse(Language.EN);
+        var version = getAssetVersion(registerResourceMessage);
+        var keywords = getAssetKeywords(registerResourceMessage)
+                .stream()
+                .map(k -> new TypedLiteral(k, language.name()))
+                .toList();
+        var mediaType = getAssetMediaType(registerResourceMessage);
+        var publisher = getAssetPublisher(registerResourceMessage);
+        var standardLicense = getAssetStandardLicense(registerResourceMessage);
+        var endpointDocumentation = getAssetEndpointDocumentation(registerResourceMessage);
+        var transportMode = getAssetTransportMode(registerResourceMessage);
+        var dataCategory = getAssetDataCategory(registerResourceMessage);
+        var dataSubcategory = getAssetDataSubcategory(registerResourceMessage);
+        var dataModel = getAssetDataModel(registerResourceMessage);
+        var geoReferenceMethod = getAssetGeoReferenceMethod(registerResourceMessage);
 
         var resource = new ResourceBuilder(registerResourceMessage.affectedResourceUri())
-                ._title_(new TypedLiteral(assetTitle, "en"))
-                ._description_(new TypedLiteral(assetDescription, "en"))
-                ._language_(Language.EN)
+                ._title_(new TypedLiteral(assetTitle, language.name()))
+                ._description_(new TypedLiteral(assetDescription, language.name()))
+                ._language_(language)
+                ._version_(version)
+                ._keyword_(keywords)
+                ._publisher_(URI.create(publisher))
+                ._standardLicense_(URI.create(standardLicense))
+                ._representation_(new RepresentationBuilder()
+                        ._language_(language)
+                        ._mediaType_(new IANAMediaTypeBuilder()._filenameExtension_(mediaType).build())
+                        .build())
+                ._resourceEndpoint_(List.of(new ConnectorEndpointBuilder()
+                        ._endpointDocumentation_(URI.create(endpointDocumentation))
+                        ._accessURL_(registerResourceMessage.affectedResourceUri())
+                        .build()))
                 .build();
-        return objectMapper.writeValueAsString(resource);
+
+        var json = (ObjectNode) objectMapper.readTree(objectMapper.writeValueAsString(resource));
+        json.set("http://w3id.org/mds#transportMode", buildTransportModeJson(transportMode));
+        json.set("http://w3id.org/mds#dataCategory", buildDataCategoryJson(dataCategory));
+        json.set("http://w3id.org/mds#dataSubcategory", buildDataSubcategoryJson(dataSubcategory));
+        json.set("http://w3id.org/mds#dataModel", buildDataModelJson(dataModel));
+        json.set("http://w3id.org/mds#geoReferenceMethod", buildGeoReferenceMethodJson(geoReferenceMethod));
+        return objectMapper.writeValueAsString(json);
+    }
+
+    private ObjectNode buildGeoReferenceMethodJson(String geoReferenceMethod) {
+        return buildStringProperty(geoReferenceMethod);
+    }
+
+    private ObjectNode buildDataModelJson(String dataModel) {
+        return buildStringProperty(dataModel);
+    }
+
+    private ObjectNode buildDataSubcategoryJson(String dataSubcategory) {
+        return buildStringProperty(dataSubcategory);
+    }
+
+    private ObjectNode buildTransportModeJson(String transportMode) {
+        return buildStringProperty(transportMode);
+    }
+
+    private ObjectNode buildDataCategoryJson(String dataCategory) {
+        return buildStringProperty(dataCategory);
+    }
+
+    private ObjectNode buildStringProperty(String property) {
+        var json = objectMapper.createObjectNode();
+        json.put("@value", property);
+        json.put("@type", "http://www.w3.org/2001/XMLSchema#string");
+        return json;
+    }
+
+    private String getAssetGeoReferenceMethod(RegisterResourceMessage registerResourceMessage) {
+        return getAssetProperty(registerResourceMessage, "http://w3id.org/mds#geoReferenceMethod");
+    }
+
+    private String getAssetDataModel(RegisterResourceMessage registerResourceMessage) {
+        return getAssetProperty(registerResourceMessage, "http://w3id.org/mds#dataModel");
+    }
+
+    private String getAssetDataSubcategory(RegisterResourceMessage registerResourceMessage) {
+        return getAssetProperty(registerResourceMessage, "http://w3id.org/mds#dataSubcategory");
+    }
+
+    private String getAssetDataCategory(RegisterResourceMessage registerResourceMessage) {
+        return getAssetProperty(registerResourceMessage, "http://w3id.org/mds#dataCategory");
+    }
+
+    private String getAssetTransportMode(RegisterResourceMessage registerResourceMessage) {
+        return getAssetProperty(registerResourceMessage, "http://w3id.org/mds#transportMode");
+    }
+
+    private String getAssetEndpointDocumentation(RegisterResourceMessage registerResourceMessage) {
+        return getAssetProperty(registerResourceMessage, "asset:prop:endpointDocumentation");
+    }
+
+    private String getAssetStandardLicense(RegisterResourceMessage registerResourceMessage) {
+        return getAssetProperty(registerResourceMessage, "asset:prop:standardLicense");
+    }
+
+    private String getAssetPublisher(RegisterResourceMessage registerResourceMessage) {
+        return getAssetProperty(registerResourceMessage, "asset:prop:publisher");
+    }
+
+    private String getAssetLanguage(RegisterResourceMessage registerResourceMessage) {
+        return getAssetProperty(registerResourceMessage, "asset:prop:language");
+    }
+
+    private static List<String> getAssetKeywords(RegisterResourceMessage registerResourceMessage) {
+        var keywords = getAssetProperty(registerResourceMessage, "asset:prop:keywords");
+        return new ArrayList<>(Arrays.asList(keywords.split(",")));
+    }
+
+    private static String getAssetMediaType(RegisterResourceMessage registerResourceMessage) {
+        return getAssetProperty(registerResourceMessage, "asset:prop:contenttype");
     }
 
     private static String getAssetDescription(RegisterResourceMessage registerResourceMessage) {
-        var assetDescription = "";
-        if (registerResourceMessage.asset().getProperties().get("asset:prop:description") != null) {
-            assetDescription = registerResourceMessage.asset().getProperties().get("asset:prop:description").toString();
-        }
-        return assetDescription;
+        return getAssetProperty(registerResourceMessage, "asset:prop:description");
     }
 
     private static String getAssetTitle(RegisterResourceMessage registerResourceMessage) {
-        var assetTitle = "";
-        if (registerResourceMessage.asset().getProperties().get("asset:prop:name") != null) {
-            assetTitle = registerResourceMessage.asset().getProperties().get("asset:prop:name").toString();
+        return getAssetProperty(registerResourceMessage, "asset:prop:name");
+    }
+
+    private static String getAssetVersion(RegisterResourceMessage registerResourceMessage) {
+        return getAssetProperty(registerResourceMessage, "asset:prop:version");
+    }
+
+    private static String getAssetProperty(RegisterResourceMessage registerResourceMessage, String property) {
+        var result = "";
+        if (checkPropertyExists(registerResourceMessage, property)) {
+            result = registerResourceMessage.asset().getProperties().get(property).toString();
         }
-        return assetTitle;
+        return result;
+    }
+
+    private static boolean checkPropertyExists(RegisterResourceMessage registerResourceMessage, String property) {
+        return registerResourceMessage.asset().getProperties().get(property) != null;
     }
 
     @Override